requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
contextlib2==21.6.0
cryptography==45.0.4
Django==5.2.3
django-cors-headers==4.7.0
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
injector==0.22.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
